# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Interview Backend API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:18080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/job-seeker/resumes/{resumeId}/publication": {
      "patch": {
        "tags": [
          "job-seeker-publication-controller"
        ],
        "operationId": "updateResumePublication",
        "parameters": [
          {
            "name": "resumeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePublicationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/job-seeker/profile/publication": {
      "patch": {
        "tags": [
          "job-seeker-publication-controller"
        ],
        "operationId": "updateProfilePublication",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePublicationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/job-seeker/portfolios/{portfolioId}/publication": {
      "patch": {
        "tags": [
          "job-seeker-publication-controller"
        ],
        "operationId": "updatePortfolioPublication",
        "parameters": [
          {
            "name": "portfolioId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePublicationResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PublicationRequest": {
        "type": "object",
        "properties": {
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE",
              "HIDDEN"
            ]
          }
        },
        "required": [
          "visibility"
        ]
      },
      "ApiResponsePublicationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PublicationResponse"
          }
        }
      },
      "PublicationResponse": {
        "type": "object",
        "properties": {
          "resourceType": {
            "type": "string"
          },
          "resourceId": {
            "type": "integer",
            "format": "int64"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE",
              "HIDDEN"
            ]
          },
          "publicProfileSlug": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}